HW 7 Completed
