+++ students/Dmitry Khlyupin/project/addToBasket.json
@@ -1,0 +1,3 @@
+{
+    "result": 1
+  }
