+++ students/Dmitry Khlyupin/project/cart.json
@@ -1,0 +1,1 @@
+{"amount":51600,"countGoods":2,"contents":[{"id_product":123,"product_name":"Ноутбук","price":45600,"quantity":1},{"id_product":3009,"product_name":"Куртка","price":3500,"quantity":1},{"id_product":5173,"product_name":"Ботинки","price":2500,"quantity":1}]}
