--- students/Dmitry Khlyupin/project/less/__variables.less
@@ -1,6 +1,0 @@
-//out: false
-
-@mainOrange: #ef5221;
-@mainWhite: #ffffff;
-@lightGrey: #eceae7;
-@darkDark: #190b09;
