+++ students/Dmitry Khlyupin/project/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "dmhlupin.github.io",
+  "version": "1.0.0",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node index"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/dmhlupin/dmhlupin.github.io.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/dmhlupin/dmhlupin.github.io/issues"
+  },
+  "homepage": "https://github.com/dmhlupin/dmhlupin.github.io#readme",
+  "dependencies": {
+    "cors": "^2.8.5",
+    "express": "^4.17.1"
+  },
+  "devDependencies": {},
+  "description": ""
+}
