+++ students/Dmitry Khlyupin/project/rest.md
@@ -1,0 +1,11 @@
+# RESTFUL API CRUD for users
+
+...
+    READ - GET /api/users?page=5 - массив пользователей
+    READ - GET /api/users/:id - конкретный пользователь с id = :id
+    READ - GET /api/users/:id/liked-books - список books пользователя с :id
+    UPDATE - PATCH||PUT /api/users/:id - обновление user с id = :id
+    CREATE - POST /api/users - создание нового пользователя
+    DELETE - POST /api/users/:id - удаление пользователя c id = :id
+...
+
