+++ students/Dmitry Khlyupin/project/static/index.html
@@ -1,0 +1,54 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>eShop</title>
+    <script src="https://cdn.jsdelivr.net/npm/vue@2/dist/vue.js" defer></script>
+    <script src="script.js" defer></script>
+    <link rel="stylesheet" href="./css/style.css">
+    <link rel="preconnect" href="https://fonts.gstatic.com">
+    <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@600&display=swap" rel="stylesheet">
+    
+    <script src="https://kit.fontawesome.com/422f9eb0c9.js" crossorigin="anonymous"></script>
+
+    <!-- <link rel="stylesheet" href="./fontawesome/css/all.css"> -->
+</head>
+<body>
+    <div id="app">
+        <header>
+            <div class="logo"><a href="#">eShop</a> </div>
+            <button v-if="filteredGoods.length < goods.length" class="clearFilterBtn" v-on:click="clearFilter">Очистить поиск</button>
+            <search  @start-search = "filterGoods" :sline="searchLine"></search>
+            <div class="card-button" v-bind:class="{ cardButtonActive: cartVisibility }" v-on:click="cartVisibility = !cartVisibility">
+                <span>Корзина </span>
+                <i v-if="!cartVisibility" class="fas fa-angle-double-down"></i>
+                <i v-else class="fas fa-angle-double-up"></i>
+            </div>
+
+        </header>
+        <main>
+            <div class="closeCartButton" v-if="cartVisibility" v-on:click="cartVisibility=!cartVisibility"><i class="fas fa-times"></i></div>
+            <cart-items 
+                v-if='cartVisibility'
+                :cart="cartGoods"
+                :full="fullPrice"
+                @remove="removeItem"
+                @add="addItem"                
+                >
+            </cart-items>
+            <goods-list 
+                v-if="filteredGoods.length > 0"
+                v-if="connected"
+                :fgoods="filteredGoods"
+                @add="addItem"
+                >
+            </goods-list>
+            <goods-error v-else-if="!connected"></goods-error>
+            <goods-empty v-else></goods-empty>
+        </main>
+    </div>
+</body>
+</html>
