+++ students/Dmitry Khlyupin/project/static/less/__variables.less
@@ -1,0 +1,6 @@
+//out: false
+
+@mainOrange: #ef5221;
+@mainWhite: #ffffff;
+@lightGrey: #eceae7;
+@darkDark: #190b09;
