+++ students/Dmitry Khlyupin/project/static/less/style.less
@@ -1,0 +1,328 @@
+//out: ../css/style.css
+
+@import "__variables.less";
+@import "__mixins.less";
+
+* {
+    padding: 0;
+    margin: 0;
+}
+
+body {
+    font-family: 'Montserrat', sans-serif;
+    position: relative;
+}
+
+a {
+    text-decoration: none;
+    color: @mainWhite;
+}
+#app {
+    width: 100%;
+    
+}
+header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    background-color: @mainOrange;
+    height: 60px;
+    padding: 0 10%;
+    box-shadow: 0px 5px 9px 0px rgba(0, 0, 0, 0.39);
+    position: fixed;
+    left: 0;
+    z-index: 100;
+    width: 80%;
+    margin: 0;
+}
+
+
+.searchContainer {
+    flex-grow: 1;
+    display: flex;
+    justify-content: flex-end;
+    margin: 0 30px;
+}
+
+.goods-search {
+    padding: 0 15px;
+    border-radius: 15px 0 0 15px;
+    height: 32px;
+    border: none;
+    font-family: 'Montserrat', sans-serif;
+    font-size: 14px;
+}
+
+.goods-search:focus,
+.goods-search:active {
+    outline: none;
+}
+
+main {
+    padding: 60px 10% 30px;
+    position:relative;
+}
+
+.logo {
+    font-size: 30px;
+    font-weight: bold;
+    color: @mainWhite;
+    margin-right: 50px;
+
+}
+
+.logo::first-letter {
+    color: @darkDark;
+    transform: rotate(45deg);
+
+}
+
+.card-button,
+.addToCart,
+.search-button {
+    background-color: #ef5221;
+    border: 1px solid #ffffff;
+    font-family: 'Montserrat', sans-serif;
+    color: #ffffff;
+    padding: 0px 15px;
+    font-size: 14px;
+    height: 30px;
+    border-radius: 15px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    outline: none;
+}
+
+.card-button span {
+    margin-right: 10px;
+}
+
+.search-button {
+    border-radius: 0 15px 15px 0;
+    height: 32px;
+}
+
+.search-button::before {
+    display: block;
+    content: '\f002';
+    font-family: 'Font Awesome 5 Free', 'Font Awesome 5 Brands';
+    font-weight: 900;
+    padding-right: 10px;
+}
+
+
+.card-button::before {
+    display: block;
+    content: '\f07a';
+    font-family: 'Font Awesome 5 Free', 'Font Awesome 5 Brands';
+    font-weight: 900;
+    padding-right: 10px;
+}
+
+.card-button:hover,
+.addToCart:hover,
+.search-button:hover {
+    background-color: #ffffff;
+    color: #ef5221;
+    border: 1px solid @mainOrange;
+    cursor: pointer;
+}
+
+.card-button:active,
+.addToCart:active,
+.search-button:active {
+    background-color: #eceae7;
+}
+
+.cardButtonActive {
+    color: @mainOrange;
+    background-color: #eceae7;
+    border: 1px solid @mainOrange;
+    font-family: 'Montserrat', sans-serif;
+    color: @mainOrange;
+    padding: 0px 15px;
+    font-size: 14px;
+    height: 30px;
+    border-radius: 15px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    outline: none;
+}
+
+.card-button:active,
+.card-button:hover {
+    outline: none;
+}
+
+.goods-list {
+
+    display: flex;
+    justify-content: space-evenly;
+    flex-wrap: wrap;
+}
+
+.goods-item {
+    padding: 20px;
+    margin: 20px 10px;
+    border: 1px solid @mainOrange;
+    border-radius: 5px;
+    width: 20%;
+    min-width: 300px;
+    height: 400px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.itemImg {
+    min-width: 200px;
+    max-width: 94%;
+    max-height: 70%;
+    border: 1px solid #eceae7;
+    box-shadow: 1px 1px 8px 1px rgba(0, 0, 0, 0.21);
+}
+.closeCartButton{
+    background-color: @mainOrange;
+    color: @mainWhite;
+    width:30px;
+    height: 30px;
+    border-radius: 50%;
+    position: absolute;
+    top: 90px;
+    right: 260px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    z-index: 101;
+    cursor: pointer;
+}
+.cartSection {
+    position: absolute;
+    right:240px;
+    top: 70px;
+    width: 600px;
+    display: flex;
+    flex-direction: column;
+    align-items: baseline;
+    background-color: @mainWhite;
+    box-shadow: 2px 2px 8px 1px rgba(0, 0, 0, 0.21);
+    border-radius: 5px;
+    padding: 20px;
+}
+
+.cartItem {
+    width: 94%;
+    display: grid;
+    grid-template-columns: 1fr 3fr 1fr 1fr 1fr;
+    border: 1px solid @lightGrey;
+    color: black;
+    font-size: 12px;
+    margin: 5px 0;
+    border-radius: 5px;
+    padding: 0 10px;
+}
+.cartItemImg{
+    height: 50px;
+
+}
+.cartItem>h3 {
+    flex-grow: 3;
+    font-size: 12px;
+    font-weight: 700;
+    display: flex;
+    align-items: center;
+}
+
+.cartItem>p {
+    flex-grow: 1;
+    vertical-align: middle;
+    display: flex;
+    align-items: center;
+}
+
+.cartHeader {
+    color: @mainOrange;
+    margin: 10px 0;
+}
+.cartItemNaming {
+    border: none;
+}
+
+.buttonBlock{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-size: 12px;
+}
+
+.removeButton, .addButton {
+    height: 26px;
+    color: #ddd;
+    
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    box-shadow: 1px 1px 5px 1px rgba(0, 0, 0, 0.1);
+
+}
+
+.removeButton {
+    background-color: @mainOrange;
+    border-radius: 13px 0 0 13px;
+    padding: 0 7px 0 10px;
+}
+
+.addButton {
+    background-color: rgb(134, 206, 76);
+    border-radius: 0 13px 13px 0;
+    padding: 0 10px 0 7px;
+}
+
+.removeButton:hover, .addButton:hover{
+    color: #fff;
+}
+.removeButton:active, .addButton:active{
+    box-shadow: none;
+}
+
+.fullCartPrice {
+    margin: 20px 0 10px;
+}
+
+.emptyBlock {
+    width: 90%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-top: 60px;    
+}
+
+.clearFilterBtn {
+    margin: 20px 0;
+    height: 32px;
+    color: @mainWhite;
+    background-color:@mainOrange;
+    border: none;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 0 10px;
+    border-radius: 16px;
+    font-family: 'Montserrat', sans-serif;
+    box-shadow: 1px 1px 5px 1px rgba(0, 0, 0, 0.1);
+    border: 1px solid @mainOrange;
+}
+
+.clearFilterBtn:hover {
+    color: @mainOrange;
+    border: 1px solid @mainOrange;
+    background-color: @mainWhite;
+}
+
+.clearFilterBtn:active, .clearFilterBtn:focus{
+    box-shadow: none;
+    outline: none;
+}
