+++ students/Dmitry Khlyupin/project/stats.json
@@ -1,0 +1,1 @@
+[{"type":"add item","id":2074,"time":"2021-03-25T09:01:26.134Z"},{"type":"add item","id":3009,"time":"2021-03-25T09:02:04.588Z"},{"type":"add item","id":3009,"time":"2021-03-25T09:02:05.198Z"},{"type":"delete item","id":123,"time":"2021-03-25T09:02:08.846Z"},{"type":"add item","id":5041,"time":"2021-03-25T09:02:10.837Z"},{"type":"add item","id":3009,"time":"2021-03-25T09:02:11.508Z"},{"type":"add item","id":3009,"time":"2021-03-25T09:02:12.027Z"},{"type":"add item","id":2074,"time":"2021-03-25T09:03:02.406Z"},{"type":"add item","id":5041,"time":"2021-03-25T09:03:03.140Z"},{"type":"add item","id":5041,"time":"2021-03-25T09:04:16.545Z"},{"type":"add item","id":5041,"time":"2021-03-25T09:04:17.244Z"},{"type":"add item","id":456,"time":"2021-03-25T09:04:39.753Z"},{"type":"add item","id":456,"time":"2021-03-25T09:04:40.244Z"},{"type":"add item","id":123,"time":"2021-03-25T09:10:52.101Z"},{"type":"add item","id":123,"time":"2021-03-25T09:10:52.292Z"},{"type":"add item","id":5173,"time":"2021-03-25T09:11:42.560Z"}]
